# openapi.yaml
openapi: 3.1.0
info:
  title: Name a Woman API
  version: 0.1.0
  description: |
    Speed-naming game — name 100 women (or men, or people) as fast as possible, validated against Wikidata.

    ## Game Loop (WebSocket)

    The game lifecycle uses a WebSocket connection at `GET /api/game/ws`.

    ### Client → Server messages

    **Start a game:**
    ```json
    {"action": "start", "category": "women", "target_count": 100}
    ```

    **Submit a guess:**
    ```json
    {"action": "guess", "name": "Marie Curie"}
    ```

    ### Server → Client messages

    **Game started:**
    ```json
    {"type": "started", "game_id": "uuid"}
    ```

    **Accepted (exact match):**
    ```json
    {"type": "accepted", "person": {...}, "accepted_count": 42, "game_complete": false}
    ```

    **Accepted (fuzzy, edit distance ≤ 2):**
    ```json
    {"type": "accepted", "person": {...}, "corrected_from": "Beyonec", "accepted_count": 43, "game_complete": false}
    ```

    **Game complete (Nth accepted guess):**
    ```json
    {"type": "accepted", "person": {...}, "accepted_count": 100, "game_complete": true,
     "completion": {"total_time_ms": 120000, "rank": 15, "total_players": 230, "percentile": 93.5}}
    ```

    **Already guessed:**
    ```json
    {"type": "already_guessed", "person": {"display_name": "Marie Curie", ...}}
    ```

    **Wrong category:**
    ```json
    {"type": "wrong_category", "person": {"display_name": "Albert Einstein", ...}}
    ```

    **Not found:**
...
servers:
...
paths:
...
  /leaderboard:
    get:
      summary: Get leaderboard
      operationId: leaderboard
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [women, men, people]
            default: women
        - name: count
          in: query
          schema:
            type: integer
            enum: [10, 100, 1000]
            default: 100
        - name: game_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: If provided, includes your_neighborhood relative to this game
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
          description: If provided, includes paginated full_leaderboard (50 per page)
      responses:
        "200":
          description: Leaderboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardResponse"
...
components:
  schemas:
...
    LeaderboardResponse:
      type: object
      required: [top_10, bottom_10]
      properties:
        top_10:
          type: array
          items:
            $ref: "#/components/schemas/LeaderboardEntry"
        bottom_10:
          type: array
          items:
            $ref: "#/components/schemas/LeaderboardEntry"
        your_neighborhood:
          $ref: "#/components/schemas/Neighborhood"
        full_leaderboard:
          $ref: "#/components/schemas/PaginatedLeaderboard"
    LeaderboardEntry:
      type: object
      required: [rank, total_time_ms, accepted_count, category, is_you]
      properties:
        rank:
          type: integer
        total_time_ms:
          type: integer
        accepted_count:
          type: integer
        category:
          type: string
        is_you:
          type: boolean
    Neighborhood:
      type: object
      required: [above, you, below]
      properties:
        above:
          type: array
          items:
            $ref: "#/components/schemas/LeaderboardEntry"
        you:
          $ref: "#/components/schemas/LeaderboardEntry"
        below:
          type: array
          items:
            $ref: "#/components/schemas/LeaderboardEntry"
    PaginatedLeaderboard:
      type: object
      required: [entries, page, total_pages, total_entries]
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/LeaderboardEntry"
        page:
          type: integer
        total_pages:
          type: integer
        total_entries:
          type: integer
...